Extract inline tests into separate test files
Split inline #[cfg(test)] modules across 10 crates into dedicated test
files, reducing source file sizes and improving navigation. The parser
and handler crates use categorized test sub-modules under src/tests/;
smaller crates use sibling tests.rs files; parser sub-modules use
sibling _tests.rs files.

No test logic changed — purely structural moves with identical coverage.
Files with only 1 test (webui-state, css_parser) are intentionally left
inline with a comment explaining why.
