+++ crates/webui-handler/src/lib.rs
@@ -499,967 +499,4 @@ pub fn handle(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::cell::RefCell;
-    use webui_protocol::{web_ui_fragment, ConditionExpr, FragmentList, WebUIFragmentAttribute};
-    use webui_test_utils::test_json;
-
-    // A simple test writer implementation
-    struct TestWriter {
-        content: RefCell<String>,
-        ended: RefCell<bool>,
-    }
-
-    impl TestWriter {
-        fn new() -> Self {
-            Self {
-                content: RefCell::new(String::new()),
-                ended: RefCell::new(false),
-            }
-        }
-
-        fn get_content(&self) -> String {
-            self.content.borrow().clone()
-        }
-
-        fn is_ended(&self) -> bool {
-            *self.ended.borrow()
-        }
-    }
-
-    impl ResponseWriter for TestWriter {
-        fn write(&mut self, content: &str) -> Result<()> {
-            self.content.borrow_mut().push_str(content);
-            Ok(())
-        }
-
-        fn end(&mut self) -> Result<()> {
-            *self.ended.borrow_mut() = true;
-            Ok(())
-        }
-    }
-
-    #[test]
-    fn test_handle_raw() {
-        // Create a simple protocol
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::raw("Hello, WebUI!")],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({});
-
-        // Create a test writer
-        let mut writer = TestWriter::new();
-
-        // Handle the protocol
-        assert!(
-            handle(&protocol, &state, &mut writer).is_ok(),
-            "Failed to handle raw protocol"
-        );
-
-        // Check the output
-        assert_eq!(writer.get_content(), "Hello, WebUI!");
-        assert!(writer.is_ended());
-    }
-
-    #[test]
-    fn test_handle_signal() {
-        // Create a protocol with a signal
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("Hello, "),
-                    WebUIFragment::signal("name", false),
-                    WebUIFragment::raw("!"),
-                ],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"name": "WebUI"});
-
-        // Create a test writer
-        let mut writer = TestWriter::new();
-
-        // Handle the protocol
-        assert!(
-            handle(&protocol, &state, &mut writer).is_ok(),
-            "Failed to handle signal protocol"
-        );
-
-        // Check the output
-        assert_eq!(writer.get_content(), "Hello, WebUI!");
-        assert!(writer.is_ended());
-    }
-
-    #[test]
-    fn test_handle_for_loop() {
-        // Create a protocol with a for loop
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("People: "),
-                    WebUIFragment::for_loop("person", "people", "person-item"),
-                ],
-            },
-        );
-
-        fragments.insert(
-            "person-item".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::signal("person.name", false),
-                    WebUIFragment::raw(", "),
-                ],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({
-            "people": [
-                {"name": "Alice"},
-                {"name": "Bob"},
-                {"name": "Charlie"}
-            ]
-        });
-
-        // Create a test writer
-        let mut writer = TestWriter::new();
-
-        // Handle the protocol
-        assert!(
-            handle(&protocol, &state, &mut writer).is_ok(),
-            "Failed to handle for loop protocol"
-        );
-
-        // Check the output
-        assert_eq!(writer.get_content(), "People: Alice, Bob, Charlie, ");
-        assert!(writer.is_ended());
-    }
-
-    #[test]
-    fn test_handle_if_condition() {
-        // Create a protocol with an if condition
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("Status: "),
-                    WebUIFragment::if_cond(
-                        webui_protocol::ConditionExpr::identifier("isActive"),
-                        "active-content",
-                    ),
-                    WebUIFragment::raw("End"),
-                ],
-            },
-        );
-
-        fragments.insert(
-            "active-content".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::raw("Active")],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-
-        // Test with isActive = true
-        let state_true = test_json!({"isActive": true});
-        let mut writer_true = TestWriter::new();
-        assert!(
-            handle(&protocol, &state_true, &mut writer_true).is_ok(),
-            "Failed to handle if condition (true case)"
-        );
-        assert_eq!(writer_true.get_content(), "Status: ActiveEnd");
-        assert!(writer_true.is_ended());
-
-        // Test with isActive = false
-        let state_false = test_json!({"isActive": false});
-        let mut writer_false = TestWriter::new();
-        assert!(
-            handle(&protocol, &state_false, &mut writer_false).is_ok(),
-            "Failed to handle if condition (false case)"
-        );
-        assert_eq!(writer_false.get_content(), "Status: End");
-        assert!(writer_false.is_ended());
-    }
-
-    #[test]
-    fn test_handle_component() {
-        // Create a protocol with a component
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("Component: "),
-                    WebUIFragment::component("my-component"),
-                ],
-            },
-        );
-
-        fragments.insert(
-            "my-component".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::raw("<div>Component Content</div>")],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({});
-
-        // Create a test writer
-        let mut writer = TestWriter::new();
-
-        // Handle the protocol
-        assert!(
-            handle(&protocol, &state, &mut writer).is_ok(),
-            "Failed to handle component protocol"
-        );
-
-        // Check the output
-        assert_eq!(
-            writer.get_content(),
-            "Component: <div>Component Content</div>"
-        );
-        assert!(writer.is_ended());
-    }
-
-    #[test]
-    fn test_missing_fragment() {
-        // Create a protocol with a missing fragment reference
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::component("missing-component")],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({});
-
-        // Create a test writer
-        let mut writer = TestWriter::new();
-
-        // Handle the protocol
-        let result = handle(&protocol, &state, &mut writer);
-
-        // Expect an error
-        assert!(result.is_err());
-        if let Err(HandlerError::MissingFragment(fragment_id)) = result {
-            assert_eq!(fragment_id, "missing-component");
-        } else {
-            panic!("Expected MissingFragment error");
-        }
-    }
-
-    #[test]
-    fn test_missing_signal_renders_empty() {
-        // A signal referencing a field absent from state should render as empty
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("Hello, "),
-                    WebUIFragment::signal("missing_field", false),
-                    WebUIFragment::raw("!"),
-                ],
-            },
-        );
-
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({});
-
-        let mut writer = TestWriter::new();
-
-        assert!(
-            handle(&protocol, &state, &mut writer).is_ok(),
-            "Missing signal should not produce an error"
-        );
-
-        assert_eq!(writer.get_content(), "Hello, !");
-        assert!(writer.is_ended());
-    }
-
-    // ── Boolean attribute rendering tests ─────────────────────────────
-
-    #[test]
-    fn test_boolean_attr_true() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<button"),
-                    WebUIFragment::attribute_boolean(
-                        "disabled",
-                        ConditionExpr::identifier("isDisabled"),
-                    ),
-                    WebUIFragment::raw(">Click</button>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"isDisabled": true});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<button disabled>Click</button>");
-    }
-
-    #[test]
-    fn test_boolean_attr_false() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<button"),
-                    WebUIFragment::attribute_boolean(
-                        "disabled",
-                        ConditionExpr::identifier("isDisabled"),
-                    ),
-                    WebUIFragment::raw(">Click</button>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"isDisabled": false});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<button>Click</button>");
-    }
-
-    #[test]
-    fn test_boolean_attr_missing() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<input type=\"checkbox\""),
-                    WebUIFragment::attribute_boolean(
-                        "checked",
-                        ConditionExpr::identifier("checked"),
-                    ),
-                    WebUIFragment::raw(">"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<input type=\"checkbox\">");
-    }
-
-    #[test]
-    fn test_boolean_attr_multiple() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<input type=\"checkbox\""),
-                    WebUIFragment::attribute_boolean(
-                        "checked",
-                        ConditionExpr::identifier("checked"),
-                    ),
-                    WebUIFragment::attribute_boolean(
-                        "disabled",
-                        ConditionExpr::identifier("disabled"),
-                    ),
-                    WebUIFragment::raw(">"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"checked": true, "disabled": false});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<input type=\"checkbox\" checked>");
-    }
-
-    // ── Simple attribute rendering tests ──────────────────────────────
-
-    #[test]
-    fn test_attribute_with_value() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<input"),
-                    WebUIFragment::attribute("value", "inputValue"),
-                    WebUIFragment::raw(">"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"inputValue": "Hello"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<input value=\"Hello\">");
-    }
-
-    #[test]
-    fn test_attribute_with_falsy_numeric() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<div name=\"test\""),
-                    WebUIFragment::attribute("handle", "number"),
-                    WebUIFragment::raw("></div>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"number": 0});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<div name=\"test\" handle=\"0\"></div>"
-        );
-    }
-
-    // ── Template attribute rendering tests ────────────────────────────
-
-    #[test]
-    fn test_mixed_attribute_template() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<input"),
-                    WebUIFragment::attribute_template("value", "attr-1"),
-                    WebUIFragment::raw(">"),
-                ],
-            },
-        );
-        fragments.insert(
-            "attr-1".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("hello "),
-                    WebUIFragment::signal("item", false),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"item": "world"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<input value=\"hello world\">");
-    }
-
-    // ── Raw signal rendering test ─────────────────────────────────────
-
-    #[test]
-    fn test_raw_signal_not_escaped() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::signal("html", false),
-                    WebUIFragment::signal("html", true),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"html": "<strong>hi</strong>"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "&lt;strong&gt;hi&lt;&#x2F;strong&gt;<strong>hi</strong>"
-        );
-    }
-
-    // ── Nested for loop tests ─────────────────────────────────────────
-
-    #[test]
-    fn test_nested_for_loop() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<div>"),
-                    WebUIFragment::for_loop("outerItem", "outerItems", "outer"),
-                    WebUIFragment::raw("</div>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "outer".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<div>"),
-                    WebUIFragment::for_loop("innerItem", "outerItem.innerItems", "inner"),
-                    WebUIFragment::raw("</div>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "inner".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::raw("<span>Inner</span>")],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({
-            "outerItems": [
-                {"innerItems": [{"name": "A"}, {"name": "B"}]},
-                {"innerItems": [{"name": "C"}]}
-            ]
-        });
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<div><div><span>Inner</span><span>Inner</span></div><div><span>Inner</span></div></div>"
-        );
-    }
-
-    #[test]
-    fn test_nested_for_with_signals() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::for_loop("item", "items", "item-tpl")],
-            },
-        );
-        fragments.insert(
-            "item-tpl".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<p>"),
-                    WebUIFragment::signal("item.name", false),
-                    WebUIFragment::raw("</p>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"items": [{"name": "Alice"}, {"name": "Bob"}]});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "<p>Alice</p><p>Bob</p>");
-    }
-
-    #[test]
-    fn test_nested_for_with_global_state() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::for_loop("item", "items", "item-tpl")],
-            },
-        );
-        fragments.insert(
-            "item-tpl".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::signal("item.name", false),
-                    WebUIFragment::raw(" - "),
-                    WebUIFragment::signal("globalTitle", false),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"items": [{"name": "A"}, {"name": "B"}], "globalTitle": "Title"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "A - TitleB - Title");
-    }
-
-    // ── For + If state scoping tests ──────────────────────────────────
-
-    #[test]
-    fn test_if_in_for_uses_local_state() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::for_loop("item", "items", "item-tpl")],
-            },
-        );
-        fragments.insert(
-            "item-tpl".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::if_cond(
-                    ConditionExpr::identifier("item.visible"),
-                    "visible-tpl",
-                )],
-            },
-        );
-        fragments.insert(
-            "visible-tpl".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::signal("item.name", false)],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"items": [{"name": "Show", "visible": true}, {"name": "Hide", "visible": false}]});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "Show");
-    }
-
-    #[test]
-    fn test_for_if_local_overrides_global() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::for_loop("item", "items", "item-tpl")],
-            },
-        );
-        fragments.insert(
-            "item-tpl".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::if_cond(
-                    ConditionExpr::identifier("item.flag"),
-                    "show-tpl",
-                )],
-            },
-        );
-        fragments.insert(
-            "show-tpl".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::raw("yes")],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        // Global flag is true, but local item.flag is false for second item
-        let state = test_json!({"flag": true, "items": [{"flag": true}, {"flag": false}]});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(writer.get_content(), "yes");
-    }
-
-    // ── Component attribute state tests ───────────────────────────────
-
-    #[test]
-    fn test_component_attr_state_simple() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<my-comp"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: "title".into(),
-                                value: "Attribute Title".into(),
-                                attr_start: true,
-                                raw_value: true,
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("my-comp"),
-                    WebUIFragment::raw("</my-comp>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "my-comp".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<span>"),
-                    WebUIFragment::signal("title", false),
-                    WebUIFragment::raw("</span>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"title": "Global Title"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<my-comp title=\"Attribute Title\"><span>Attribute Title</span></my-comp>"
-        );
-    }
-
-    #[test]
-    fn test_component_attr_state_template() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<my-comp"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: "title".into(),
-                                template: "title-attr".into(),
-                                attr_start: true,
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("my-comp"),
-                    WebUIFragment::raw("</my-comp>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "title-attr".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("hello "),
-                    WebUIFragment::signal("item", false),
-                ],
-            },
-        );
-        fragments.insert(
-            "my-comp".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<span>"),
-                    WebUIFragment::signal("title", false),
-                    WebUIFragment::raw("</span>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"item": "<world>"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<my-comp title=\"hello &lt;world&gt;\"><span>hello &lt;world&gt;</span></my-comp>"
-        );
-    }
-
-    #[test]
-    fn test_component_attr_camel_case() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<my-comp"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: "data-title".into(),
-                                template: "dt-attr".into(),
-                                attr_start: true,
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("my-comp"),
-                    WebUIFragment::raw("</my-comp>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "dt-attr".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("prefix "),
-                    WebUIFragment::signal("item", false),
-                ],
-            },
-        );
-        fragments.insert(
-            "my-comp".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<span>"),
-                    WebUIFragment::signal("dataTitle", false),
-                    WebUIFragment::raw("</span>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"item": "a&b"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<my-comp data-title=\"prefix a&amp;b\"><span>prefix a&amp;b</span></my-comp>"
-        );
-    }
-
-    #[test]
-    fn test_component_complex_attr() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<my-comp"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: ":item".into(),
-                                value: "complexItem".into(),
-                                attr_start: true,
-                                complex: true,
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("my-comp"),
-                    WebUIFragment::raw("</my-comp>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "my-comp".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<span>"),
-                    WebUIFragment::signal("item.foo", false),
-                    WebUIFragment::raw("</span><p>"),
-                    WebUIFragment::signal("item.bar", false),
-                    WebUIFragment::raw("</p>"),
-                ],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"complexItem": {"foo": 1, "bar": "true"}});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<my-comp><span>1</span><p>true</p></my-comp>"
-        );
-    }
-
-    #[test]
-    fn test_component_no_parent_pollution() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<parent"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: "var".into(),
-                                value: "var".into(),
-                                attr_start: true,
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("parent"),
-                    WebUIFragment::raw("</parent>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "parent".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("Before: "),
-                    WebUIFragment::signal("var", false),
-                    WebUIFragment::raw("<child foo"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: "var".into(),
-                                value: "replaced".into(),
-                                raw_value: true,
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("child"),
-                    WebUIFragment::raw("Label</child>After: "),
-                    WebUIFragment::signal("var", false),
-                ],
-            },
-        );
-        fragments.insert("child".to_string(), FragmentList { fragments: vec![] });
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"var": "original"});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<parent var=\"original\">Before: original<child foo var=\"replaced\">Label</child>After: original</parent>"
-        );
-    }
-
-    #[test]
-    fn test_component_boolean_attr_state() {
-        let mut fragments = HashMap::new();
-        fragments.insert(
-            "index.html".to_string(),
-            FragmentList {
-                fragments: vec![
-                    WebUIFragment::raw("<my-comp"),
-                    WebUIFragment {
-                        fragment: Some(web_ui_fragment::Fragment::Attribute(
-                            WebUIFragmentAttribute {
-                                name: "disabled".into(),
-                                attr_start: true,
-                                condition_tree: Some(ConditionExpr::identifier("isDisabled")),
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    WebUIFragment::raw(">"),
-                    WebUIFragment::component("my-comp"),
-                    WebUIFragment::raw("</my-comp>"),
-                ],
-            },
-        );
-        fragments.insert(
-            "my-comp".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::if_cond(
-                    ConditionExpr::identifier("disabled"),
-                    "show",
-                )],
-            },
-        );
-        fragments.insert(
-            "show".to_string(),
-            FragmentList {
-                fragments: vec![WebUIFragment::raw("disabled!")],
-            },
-        );
-        let protocol = WebUIProtocol { fragments };
-        let state = test_json!({"isDisabled": true});
-        let mut writer = TestWriter::new();
-        handle(&protocol, &state, &mut writer).unwrap();
-        assert_eq!(
-            writer.get_content(),
-            "<my-comp disabled>disabled!</my-comp>"
-        );
-    }
-}
+mod tests;
