+++ crates/webui-parser/src/lib.rs
@@ -1100,1240 +1100,4 @@ impl HtmlParser {
 }
 
 #[cfg(test)]
-mod tests {
-    use webui_test_utils::*;
-
-    use super::*;
-
-    #[test]
-    fn test_parse_signal() {
-        let mut parser = HtmlParser::new();
-        let html = "Hello, {{name}}!";
-        let result = parser.parse("test.html", html);
-
-        assert!(result.is_ok());
-        let fragment_records = parser.into_fragment_records();
-
-        assert_stream!(
-            fragment_records,
-            "test.html",
-            [raw("Hello, "), signal("name"), raw("!"),]
-        );
-    }
-
-    #[test]
-    fn test_parse_raw_signal() {
-        let mut parser = HtmlParser::new();
-        let html = "Hello, {{{html_content}}}!";
-        let result = parser.parse("test.html", html);
-
-        assert!(result.is_ok());
-        let fragment_records = parser.into_fragment_records();
-
-        assert_stream!(
-            fragment_records,
-            "test.html",
-            [raw("Hello, "), signal_raw("html_content"), raw("!"),]
-        );
-    }
-
-    #[test]
-    fn test_parse_for_directive() {
-        let mut parser = HtmlParser::new();
-        let html = r#"<for each="item in items"><div class="item">{{item.name}}</div></for>"#;
-
-        let result = parser.parse("test.html", html);
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let fragment_records = parser.into_fragment_records();
-        println!("Fragment records: {:#?}", fragment_records);
-
-        assert_stream!(
-            fragment_records,
-            "test.html",
-            [for_loop("item", "items", "for-1"),]
-        );
-
-        // Verify the sub-fragment contains our item content
-        assert_stream!(
-            fragment_records,
-            "for-1",
-            [
-                raw("<div class=\"item\">"),
-                signal("item.name"),
-                raw("</div>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parse_if_directive() {
-        let mut parser = HtmlParser::new();
-        let html = r#"<if condition="isLoggedIn"><div>Welcome back, {{username}}!</div></if>"#;
-
-        let result = parser.parse("test.html", html);
-
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let fragment_records = parser.into_fragment_records();
-        println!("Fragment records: {:#?}", fragment_records);
-
-        assert_stream!(fragment_records, "test.html", [if_cond("if-1"),]);
-
-        // Verify the sub-fragment contains our content
-        assert_stream!(
-            fragment_records,
-            "if-1",
-            [
-                raw("<div>Welcome back, "),
-                signal("username"),
-                raw("!</div>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_component_directive() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component(
-                "my-component",
-                "<div>My Component</div>",
-                Some("div { color: blue; }"),
-            )
-            .expect("Failed to register component");
-
-        let result = parser.parse("test.html", "<my-component></my-component>");
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let records = parser.into_fragment_records();
-
-        assert_stream!(
-            records,
-            "test.html",
-            [
-                raw("<my-component>"),
-                component("my-component"),
-                raw("</my-component>"),
-            ]
-        );
-
-        // Component template stream should be wrapped with shadow DOM template + style link
-        let comp = &records["my-component"].fragments;
-        assert_eq!(comp.len(), 1);
-        assert!(
-            matches!(comp[0].fragment.as_ref(), Some(Fragment::Raw(raw)) if
-                raw.value.contains("<template shadowrootmode=\"open\">") && raw.value.contains("<div>My Component</div>"))
-        );
-    }
-
-    #[test]
-    fn test_component_directive_with_slots() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component(
-                "my-component",
-                "<div>My Component</div>",
-                Some("div { color: blue; }"),
-            )
-            .expect("Failed to register component");
-
-        let result = parser.parse(
-            "test.html",
-            "Hello<my-component><p>World</p></my-component>",
-        );
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let records = parser.into_fragment_records();
-        let fragments = &records["test.html"].fragments;
-
-        // Entry: raw(Hello<my-component>) + component + raw(<p>World</p></my-component>)
-        assert!(fragments.len() >= 3);
-        // First fragment should contain "Hello" and "<my-component>"
-        assert!(
-            matches!(fragments[0].fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.contains("Hello") && raw.value.contains("<my-component>"))
-        );
-        // Should have component fragment
-        assert!(fragments.iter().any(|f| matches!(
-            f.fragment.as_ref(),
-            Some(Fragment::Component(c)) if c.fragment_id == "my-component"
-        )));
-        // Should end with closing tag
-        let last = fragments.last().unwrap();
-        assert!(
-            matches!(last.fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.contains("</my-component>"))
-        );
-    }
-
-    // ── Component template wrapping tests ────────────────────────────
-
-    #[test]
-    fn test_component_no_double_wrap_template() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component(
-                "custom-element",
-                r#"<template foo="bar"><slot></slot></template>"#,
-                None,
-            )
-            .expect("register");
-        let result = parser.parse("index.html", "<custom-element>Hello</custom-element>");
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-element>"),
-                component("custom-element"),
-                raw("Hello</custom-element>"),
-            ]
-        );
-
-        assert_stream!(
-            records,
-            "custom-element",
-            [raw(r#"<template foo="bar"><slot></slot></template>"#),]
-        );
-    }
-
-    #[test]
-    fn test_component_styled_no_double_wrap() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component(
-                "custom-element",
-                r#"<template foo="bar"><slot></slot></template>"#,
-                Some("div { color: red; }"),
-            )
-            .expect("register");
-        let result = parser.parse("index.html", "<custom-element>Hello</custom-element>");
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_stream!(
-            records,
-            "custom-element",
-            [raw(
-                r#"<template foo="bar"><link rel="stylesheet" href="./custom-element.css"><slot></slot></template>"#
-            ),]
-        );
-    }
-
-    #[test]
-    fn test_component_strip_runtime_attrs() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component(
-                "custom-element",
-                r#"<template @click={foo} :bar="baz" ?bool="true"><slot></slot></template>"#,
-                None,
-            )
-            .expect("register");
-        let result = parser.parse("index.html", "<custom-element>Hello</custom-element>");
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_stream!(
-            records,
-            "custom-element",
-            [raw("<template><slot></slot></template>"),]
-        );
-    }
-
-    #[test]
-    fn test_component_with_slots_and_attrs() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component("custom-element", "<slot></slot>", None)
-            .expect("register");
-        let result = parser.parse(
-            "index.html",
-            r#"<custom-element appearance="subtle">Hello World</custom-element>"#,
-        );
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-element"),
-                attr_raw_start("appearance", "subtle"),
-                raw(">"),
-                component("custom-element"),
-                raw("Hello World</custom-element>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_component_legacy_no_styles() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component("custom-element", "<div>Custom Element</div>", None)
-            .expect("register");
-        let result = parser.parse("index.html", "<custom-element></custom-element>");
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-element>"),
-                component("custom-element"),
-                raw("</custom-element>"),
-            ]
-        );
-
-        assert_stream!(
-            records,
-            "custom-element",
-            [raw(
-                "<template shadowrootmode=\"open\"><div>Custom Element</div></template>"
-            ),]
-        );
-    }
-
-    #[test]
-    fn test_component_self_closing() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component("custom-widget", "<div>Widget Content</div>", None)
-            .expect("register");
-        let result = parser.parse("index.html", r#"<custom-widget config="{{settings}}" />"#);
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-widget"),
-                attr_start("config", "settings"),
-                raw("/>"),
-                component("custom-widget"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_component_nested_self_closing_in_slot() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component("custom-icon", "<svg><slot></slot></svg>", None)
-            .expect("register");
-        let result = parser.parse(
-            "index.html",
-            r##"<custom-icon><use href="#icon-{{iconName}}" /></custom-icon>"##,
-        );
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-icon>"),
-                component("custom-icon"),
-                raw("<use"),
-                attr_template("href", "attr-1"),
-                raw("/></custom-icon>"),
-            ]
-        );
-
-        assert_stream!(
-            records,
-            "custom-icon",
-            [raw(
-                "<template shadowrootmode=\"open\"><svg><slot></slot></svg></template>"
-            ),]
-        );
-    }
-
-    #[test]
-    fn test_component_leading_boolean_attr_start() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component("custom-element", "<slot></slot>", None)
-            .expect("register");
-        let result = parser.parse(
-            "index.html",
-            r#"<custom-element ?disabled="{{isDisabled}}" title="Hello"></custom-element>"#,
-        );
-        assert!(result.is_ok());
-        let records = parser.into_fragment_records();
-
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-element"),
-                // First dynamic attr: boolean with attrStart
-                bool_attr_start("disabled", "isDisabled"),
-                // Static attr after dynamic: rawValue
-                attr_raw("title", "Hello"),
-                raw(">"),
-                component("custom-element"),
-                raw("</custom-element>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_component_meta_link_tags() {
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<head><meta charset="utf-8" /><link rel="stylesheet" href="{{cssFile}}" /></head>"#,
-        );
-        assert!(fragments.len() >= 3);
-        assert!(
-            matches!(fragments[0].fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.contains("<head><meta charset=\"utf-8\"") && raw.value.contains("<link"))
-        );
-        assert!(
-            matches!(fragments[1].fragment.as_ref(), Some(Fragment::Attribute(a)) if a.name == "href" && a.value == "cssFile")
-        );
-        assert!(
-            matches!(fragments[2].fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.contains("/></head>"))
-        );
-    }
-
-    #[test]
-    fn test_nested_directives() {
-        let mut parser = HtmlParser::new();
-        let html = r#"<for each="category in categories">
-            <if condition="category.hasItems">
-                <for each="item in category.items">
-                   {{item.title}}
-                </for>
-            </if>
-        </for>"#;
-
-        let result = parser.parse("test.html", html);
-
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let fragment_records = parser.into_fragment_records();
-
-        assert_stream!(
-            fragment_records,
-            "test.html",
-            [for_loop("category", "categories", "for-1"),]
-        );
-
-        assert_stream!(fragment_records, "for-1", [if_cond("if-1"),]);
-
-        assert_stream!(
-            fragment_records,
-            "if-1",
-            [for_loop("item", "category.items", "for-2"),]
-        );
-
-        assert_stream!(fragment_records, "for-2", [signal("item.title"),]);
-    }
-
-    #[test]
-    fn test_complex_directives() {
-        let mut parser = HtmlParser::new();
-        let html = r#"<for each="category in categories">
-            <div class="category">
-                <h2>{{category.name}}</h2>
-                <if condition="category.hasItems">
-                    <ul>
-                        <for each="item in category.items">
-                            <li>{{item.title}}</li>
-                        </for>
-                    </ul>
-                </if>
-            </div>
-        </for>"#;
-
-        let result = parser.parse("test.html", html);
-
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let fragment_records = parser.into_fragment_records();
-
-        assert_stream!(
-            fragment_records,
-            "test.html",
-            [for_loop("category", "categories", "for-1"),]
-        );
-
-        // Verify for fragments contains the category.name signal
-        assert_stream!(
-            fragment_records,
-            "for-1",
-            [
-                raw("<div class=\"category\"><h2>"),
-                signal("category.name"),
-                raw("</h2>"),
-                if_cond("if-1"),
-                raw("</div>"),
-            ]
-        );
-
-        // Verify nested if condition.
-        assert_stream!(
-            fragment_records,
-            "if-1",
-            [
-                raw("<ul>"),
-                for_loop("item", "category.items", "for-2"),
-                raw("</ul>"),
-            ]
-        );
-
-        // Verify nested for each.
-        assert_stream!(
-            fragment_records,
-            "for-2",
-            [raw("<li>"), signal("item.title"), raw("</li>"),]
-        );
-    }
-
-    // ── Attribute fragment tests ─────────────────────────────────────────
-
-    /// Helper to parse HTML and return the fragments for the entry stream.
-    fn parse_and_get_fragments(html: &str) -> (Vec<WebUIFragment>, WebUIFragmentRecords) {
-        let mut parser = HtmlParser::new();
-        let result = parser.parse("index.html", html);
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let records = parser.into_fragment_records();
-        let fragments = records
-            .get("index.html")
-            .expect("Failed to get index.html fragment")
-            .fragments
-            .clone();
-        (fragments, records)
-    }
-
-    #[test]
-    fn test_attribute_handlebars_in_href() {
-        // Port of: 'should process handlebars from attributes as signals'
-        let (fragments, _) = parse_and_get_fragments(r#"<a href="{{url}}">{{name}}</a>"#);
-        assert_fragments!(
-            fragments,
-            [
-                raw("<a"),
-                attr("href", "url"),
-                raw(">"),
-                signal("name"),
-                raw("</a>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_boolean_with_handlebars() {
-        // Port of: 'should process boolean attribute with handlebars expression'
-        let (fragments, _) =
-            parse_and_get_fragments("<button ?disabled={{isDisabled}}>Click</button>");
-        assert_fragments!(
-            fragments,
-            [
-                raw("<button"),
-                bool_attr("disabled", "isDisabled"),
-                raw(">Click</button>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_multiple_boolean() {
-        // Port of: 'should process multiple boolean attributes'
-        // <input ?checked={{isChecked}} ?disabled={{isDisabled}} />
-        let (fragments, _) =
-            parse_and_get_fragments("<input ?checked={{isChecked}} ?disabled={{isDisabled}} />");
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<input"),
-                bool_attr("checked", "isChecked"),
-                bool_attr("disabled", "isDisabled"),
-                raw("/>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_boolean_and_regular_together() {
-        // Port of: 'should process a boolean attribute and a regular attribute together'
-        // <input ?checked="{{isChecked}}" type="checkbox">Hi</input>
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<input ?checked="{{isChecked}}" type="checkbox">Hi</input>"#,
-        );
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<input"),
-                bool_attr("checked", "isChecked"),
-                raw(" type=\"checkbox\">Hi</input>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_boolean_sandwiched() {
-        // Port of: 'should process a boolean attribute sandwiched between regular attributes'
-        // <input version={{edition}} ?checked="{{isChecked}}" type="checkbox">Hi</input>
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<input version={{edition}} ?checked="{{isChecked}}" type="checkbox">Hi</input>"#,
-        );
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<input"),
-                attr("version", "edition"),
-                bool_attr("checked", "isChecked"),
-                raw(" type=\"checkbox\">Hi</input>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_boolean_ending() {
-        // Port of: 'should process html ending with boolean attribute correctly'
-        // <input version={{edition}} ?checked="{{isChecked}}">Hi</input>
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<input version={{edition}} ?checked="{{isChecked}}">Hi</input>"#,
-        );
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<input"),
-                attr("version", "edition"),
-                bool_attr("checked", "isChecked"),
-                raw(">Hi</input>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_boolean_dotted_path() {
-        // Port of: 'should process boolean attribute with dotted path'
-        // <div ?checked={{layout.isPinned}}>Content</div>
-        let (fragments, _) =
-            parse_and_get_fragments("<div ?checked={{layout.isPinned}}>Content</div>");
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<div"),
-                bool_attr("checked", "layout.isPinned"),
-                raw(">Content</div>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_colon_prefixed_complex() {
-        // Port of: 'should process colon-prefixed attribute with handlebars'
-        // <my-component :config="{{settings}}"></my-component>
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<my-component :config="{{settings}}"></my-component>"#);
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<my-component"),
-                attr_complex(":config", "settings"),
-                raw("></my-component>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_multiple_colon_prefixed() {
-        // Port of: 'should process multiple colon-prefixed complex attributes'
-        // <my-component :prop1="{{val1}}" :prop2="{{val2}}"></my-component>
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<my-component :prop1="{{val1}}" :prop2="{{val2}}"></my-component>"#,
-        );
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<my-component"),
-                attr_complex(":prop1", "val1"),
-                attr_complex(":prop2", "val2"),
-                raw("></my-component>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_mixed_normal_boolean_colon() {
-        // Port of: 'should process mixed normal, boolean, and colon-prefixed attributes'
-        // <my-component id="comp" :config="{{settings}}" ?enabled="{{isEnabled}}"></my-component>
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<my-component id="comp" :config="{{settings}}" ?enabled="{{isEnabled}}"></my-component>"#,
-        );
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<my-component id=\"comp\""),
-                attr_complex(":config", "settings"),
-                bool_attr("enabled", "isEnabled"),
-                raw("></my-component>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_attribute_reject_boolean_without_handlebars() {
-        // Port of: 'should reject boolean attribute without handlebars'
-        // <input ?checked="name"></input>
-        let (fragments, _) = parse_and_get_fragments(r#"<input ?checked="name"></input>"#);
-
-        // Boolean attribute is silently dropped
-        assert_fragments!(fragments, [raw("<input></input>"),]);
-    }
-
-    #[test]
-    fn test_attribute_reject_boolean_with_partial_handlebars() {
-        // Port of: 'should reject boolean attribute with partial handlebars'
-        // <input ?checked="Hello {{name}}"></input>
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<input ?checked="Hello {{name}}"></input>"#);
-
-        // Boolean attribute is silently dropped
-        assert_fragments!(fragments, [raw("<input></input>"),]);
-    }
-
-    #[test]
-    fn test_attribute_reject_boolean_with_plain_value() {
-        // Port of: 'should reject boolean attribute with plain value'
-        // <button ?disabled="true">Click</button>
-        let (fragments, _) = parse_and_get_fragments(r#"<button ?disabled="true">Click</button>"#);
-
-        // Boolean attribute is silently dropped
-        assert_fragments!(fragments, [raw("<button>Click</button>"),]);
-    }
-
-    #[test]
-    fn test_attribute_mixed_static_dynamic() {
-        // Port of: 'should process mixed attributes correctly'
-        // <input value="hello {{world}}">Hi</input>
-        let (fragments, records) =
-            parse_and_get_fragments(r#"<input value="hello {{world}}">Hi</input>"#);
-
-        assert_fragments!(
-            fragments,
-            [
-                raw("<input"),
-                attr_template("value", "attr-1"),
-                raw(">Hi</input>"),
-            ]
-        );
-
-        // Verify the template sub-stream
-        assert_stream!(records, "attr-1", [raw("hello "), signal("world"),]);
-    }
-
-    // ── Body signal tests ─────────────────────────────────────────────
-
-    #[test]
-    fn test_body_signals() {
-        let (fragments, _) = parse_and_get_fragments("<body><app-shell></app-shell></body>");
-        assert_fragments!(
-            fragments,
-            [
-                raw("<body>"),
-                signal_raw("body_start"),
-                raw("<app-shell></app-shell>"),
-                signal_raw("body_end"),
-                raw("</body>"),
-            ]
-        );
-    }
-
-    // ── Empty for handling tests ──────────────────────────────────────
-
-    #[test]
-    fn test_empty_for_produces_nothing() {
-        let (fragments, records) =
-            parse_and_get_fragments(r#"<div><for each="item in items"></for></div>"#);
-        assert_fragments!(fragments, [raw("<div></div>"),]);
-        assert!(!records.contains_key("for-1"));
-    }
-
-    // ── Self-closing / void element tests ─────────────────────────────
-
-    #[test]
-    fn test_self_closing_svg_path() {
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<svg width="19"><path d="foo" fill="currentcolor"/></svg>"#);
-        assert_fragments!(
-            fragments,
-            [raw(
-                r#"<svg width="19"><path d="foo" fill="currentcolor"/></svg>"#
-            ),]
-        );
-    }
-
-    #[test]
-    fn test_html5_void_elements() {
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<div><img src="test.jpg" alt="test"><br><hr><input type="text"></div>"#,
-        );
-        assert_fragments!(
-            fragments,
-            [raw(
-                r#"<div><img src="test.jpg" alt="test"><br><hr><input type="text"></div>"#
-            ),]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_with_dynamic_attributes() {
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<img src="{{imageUrl}}" alt="{{imageAlt}}" />"#);
-        assert_fragments!(
-            fragments,
-            [
-                raw("<img"),
-                attr("src", "imageUrl"),
-                attr("alt", "imageAlt"),
-                raw("/>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_with_boolean_attributes() {
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<input type="checkbox" ?checked="{{isSelected}}" ?disabled="{{isDisabled}}" />"#,
-        );
-        assert_fragments!(
-            fragments,
-            [
-                raw("<input type=\"checkbox\""),
-                bool_attr("checked", "isSelected"),
-                bool_attr("disabled", "isDisabled"),
-                raw("/>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_multiple_self_closing_in_sequence() {
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<img src="1.jpg" /><br /><img src="2.jpg" />"#);
-        assert_fragments!(
-            fragments,
-            [raw(r#"<img src="1.jpg"/><br/><img src="2.jpg"/>"#),]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_with_mixed_content() {
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<div>Text before<img src="{{url}}" />Text after</div>"#);
-        assert_fragments!(
-            fragments,
-            [
-                raw("<div>Text before<img"),
-                attr("src", "url"),
-                raw("/>Text after</div>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_svg_elements() {
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<svg><circle cx="{{x}}" cy="{{y}}" r="5" /><rect width="10" height="10" /></svg>"#,
-        );
-        assert_fragments!(
-            fragments,
-            [
-                raw("<svg><circle"),
-                attr("cx", "x"),
-                attr("cy", "y"),
-                raw(r#" r="5"/><rect width="10" height="10"/></svg>"#),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_inside_for_loop() {
-        let (fragments, records) = parse_and_get_fragments(
-            r#"<for each="item in items"><img src="{{item.url}}" /></for>"#,
-        );
-        assert_fragments!(fragments, [for_loop("item", "items", "for-1"),]);
-        assert_stream!(
-            records,
-            "for-1",
-            [raw("<img"), attr("src", "item.url"), raw("/>"),]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_whitespace_variations() {
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<img src="test.jpg"/><input type="text" /><br/>"#);
-        assert_fragments!(
-            fragments,
-            [raw(r#"<img src="test.jpg"/><input type="text"/><br/>"#),]
-        );
-    }
-
-    #[test]
-    fn test_deeply_nested_self_closing() {
-        let (fragments, _) = parse_and_get_fragments(
-            r#"<div><section><article><img src="deep.jpg" /><br /></article></section></div>"#,
-        );
-        assert_fragments!(
-            fragments,
-            [raw(
-                r#"<div><section><article><img src="deep.jpg"/><br/></article></section></div>"#
-            ),]
-        );
-    }
-
-    #[test]
-    fn test_self_closing_vs_empty_regular_tags() {
-        let (fragments, _) =
-            parse_and_get_fragments(r#"<div></div><img src="test.jpg" /><span></span>"#);
-        assert_fragments!(
-            fragments,
-            [raw(r#"<div></div><img src="test.jpg"/><span></span>"#),]
-        );
-    }
-
-    // ── Feature 1: Custom template attribute on <for> ────────────────────
-
-    #[test]
-    fn test_for_custom_template_attribute() {
-        // Port of: 'should process transient node for with template'
-        let (fragments, records) = parse_and_get_fragments(
-            r#"<for each="item in items" template="static"><span>Item</span></for>"#,
-        );
-        assert_fragments!(fragments, [for_loop("item", "items", "static"),]);
-        assert_stream!(records, "static", [raw("<span>Item</span>"),]);
-    }
-
-    #[test]
-    fn test_for_recursive_template() {
-        // Port of: 'should process recursive transient nodes'
-        let mut parser = HtmlParser::new();
-        let html = r#"<for template="static" each="outerItem in outerItems"><div><span>{{outerItem.name}}</span><for template="static" each="innerItem in innerItems" /></div></for>"#;
-        let result = parser.parse("index.html", html);
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let records = parser.into_fragment_records();
-
-        assert_fragments!(
-            records["index.html"].fragments,
-            [for_loop("outerItem", "outerItems", "static"),]
-        );
-
-        assert_stream!(
-            records,
-            "static",
-            [
-                raw("<div><span>"),
-                signal("outerItem.name"),
-                raw("</span>"),
-                for_loop("innerItem", "innerItems", "static"),
-                raw("</div>"),
-            ]
-        );
-    }
-
-    // ── Feature 2: <if> / <for> with multiple children ──────────────────
-
-    #[test]
-    fn test_if_multiple_children() {
-        // Port of: 'should handle <if> with multiple children'
-        let (fragments, records) =
-            parse_and_get_fragments(r#"<if condition="valid"><p>hello</p><p>world</p></if>"#);
-        assert_fragments!(fragments, [if_cond("if-1"),]);
-        assert_stream!(records, "if-1", [raw("<p>hello</p><p>world</p>"),]);
-    }
-
-    #[test]
-    fn test_for_multiple_children() {
-        // Port of: 'should handle <for> with multiple children'
-        let (fragments, records) =
-            parse_and_get_fragments(r#"<for each="item in items"><p>hello</p><p>world</p></for>"#);
-        assert_fragments!(fragments, [for_loop("item", "items", "for-1"),]);
-        assert_stream!(records, "for-1", [raw("<p>hello</p><p>world</p>"),]);
-    }
-
-    // ── Feature 3: Handlebars at beginning/end of text ──────────────────
-
-    #[test]
-    fn test_handlebars_at_beginning() {
-        // Port of: 'should process handlebars from text at beginning'
-        let (fragments, _) = parse_and_get_fragments("{{first}}");
-        assert_fragments!(fragments, [signal("first"),]);
-    }
-
-    #[test]
-    fn test_handlebars_at_beginning_and_raw() {
-        // Port of: 'should process handlebars from text at beginning and raw'
-        let (fragments, _) = parse_and_get_fragments("{{first}}test");
-        assert_fragments!(fragments, [signal("first"), raw("test"),]);
-    }
-
-    #[test]
-    fn test_handlebars_raw_and_end() {
-        // Port of: 'should process handlebars from text at raw and end'
-        let (fragments, _) = parse_and_get_fragments("test{{first}}");
-        assert_fragments!(fragments, [raw("test"), signal("first"),]);
-    }
-
-    // ── Feature 4: Handlebars edge cases ────────────────────────────────
-
-    #[test]
-    fn test_handlebars_invalid_triple_open() {
-        // Port of: 'should not process handlebars when invalid'
-        let (fragments, _) = parse_and_get_fragments("{{{invalid}}");
-        assert_fragments!(fragments, [raw("{{{invalid}}"),]);
-    }
-
-    #[test]
-    fn test_handlebars_four_open_braces() {
-        // Port of: 'should not process handlebars when invalid since triple exists'
-        let (fragments, _) = parse_and_get_fragments("{{{{invalid}}");
-        assert_fragments!(fragments, [raw("{{{{invalid}}"),]);
-    }
-
-    #[test]
-    fn test_handlebars_five_open_with_valid_double() {
-        // Port of: 'should not process handlebars when invalid but with valid triple'
-        let (fragments, _) = parse_and_get_fragments("{{{{{invalid}}");
-        assert_fragments!(fragments, [raw("{{{"), signal("invalid"),]);
-    }
-
-    #[test]
-    fn test_entities_preserved() {
-        // Port of: 'should process entities correctly'
-        let (fragments, _) = parse_and_get_fragments("<p>Hello&#125;World</p>");
-        assert_fragments!(fragments, [raw("<p>Hello&#125;World</p>"),]);
-    }
-
-    // ── Feature 5: DOCTYPE handling ─────────────────────────────────────
-
-    #[test]
-    fn test_doctype_preserved() {
-        // DOCTYPE should be preserved as raw content
-        let (fragments, _) = parse_and_get_fragments("<!DOCTYPE html><html><head></head></html>");
-        assert!(fragments.len() >= 1);
-        assert!(
-            matches!(fragments[0].fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.contains("<!DOCTYPE html>"))
-        );
-    }
-
-    // ── Feature 6: Component attribute skip / multiple nested ───────────
-
-    #[test]
-    fn test_component_attr_skip() {
-        // Port of: 'should set attrSkip for skipped component attributes'
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component("custom-element", "<slot></slot>", None)
-            .expect("register");
-        let html = r#"<custom-element :config="{{config}}" class="{{value0}}" style="{{value1}}" role="{{value2}}" data-test="{{value3}}" aria-test="{{value4}}"></custom-element>"#;
-        let result = parser.parse("index.html", html);
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let records = parser.into_fragment_records();
-
-        // <custom-element, :config(attrStart), class(attrSkip), style(attrSkip),
-        // role(attrSkip), data-test(attrSkip), aria-test(attrSkip), >, component, </custom-element>
-        assert_fragments!(
-            records["index.html"].fragments,
-            [
-                raw("<custom-element"),
-                // :config with attrStart
-                attr_complex_start(":config", "config"),
-                // Skipped attrs
-                attr_skip("class", "value0"),
-                attr_skip("style", "value1"),
-                attr_skip("role", "value2"),
-                attr_skip("data-test", "value3"),
-                attr_skip("aria-test", "value4"),
-                raw(">"),
-                component("custom-element"),
-                raw("</custom-element>"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_component_multiple_nested() {
-        // Port of: 'handle multiple nested web components'
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry
-            .register_component(
-                "custom-element",
-                "<custom-child></custom-child><slot></slot>",
-                None,
-            )
-            .expect("register");
-        parser
-            .component_registry
-            .register_component("custom-button", "<slot></slot>", None)
-            .expect("register");
-        parser
-            .component_registry
-            .register_component("custom-child", "<h1>Hello World!</h1>", None)
-            .expect("register");
-
-        let html = r#"<for each="item in items"><custom-element><custom-button>Ok</custom-button></custom-element></for>"#;
-        let result = parser.parse("index.html", html);
-        assert!(result.is_ok(), "Parse error: {:?}", result.err());
-        let records = parser.into_fragment_records();
-
-        // Entry stream
-        assert_fragments!(
-            records["index.html"].fragments,
-            [for_loop("item", "items", "for-1"),]
-        );
-
-        // For stream
-        assert_stream!(
-            records,
-            "for-1",
-            [
-                raw("<custom-element>"),
-                component("custom-element"),
-                raw("<custom-button>"),
-                component("custom-button"),
-                raw("Ok</custom-button></custom-element>"),
-            ]
-        );
-
-        // Component streams — custom-element has contains() checks, keep manual
-        let ce = &records["custom-element"].fragments;
-        assert_eq!(ce.len(), 3);
-        assert!(
-            matches!(ce[0].fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.starts_with("<template shadowrootmode=\"open\"><custom-child>"))
-        );
-        assert!(
-            matches!(ce[1].fragment.as_ref(), Some(Fragment::Component(c)) if c.fragment_id == "custom-child")
-        );
-        assert!(
-            matches!(ce[2].fragment.as_ref(), Some(Fragment::Raw(raw)) if raw.value.contains("</custom-child><slot></slot></template>"))
-        );
-
-        assert_stream!(
-            records,
-            "custom-button",
-            [raw(
-                "<template shadowrootmode=\"open\"><slot></slot></template>"
-            ),]
-        );
-
-        assert_stream!(
-            records,
-            "custom-child",
-            [raw(
-                "<template shadowrootmode=\"open\"><h1>Hello World!</h1></template>"
-            ),]
-        );
-    }
-
-    // ── Error handling tests ──────────────────────────────────────────
-
-    #[test]
-    fn test_invalid_markup_returns_error() {
-        // Port of: 'should fail with invalid markup'
-        // tree-sitter is lenient — it recovers from unclosed tags
-        let mut parser = HtmlParser::new();
-        let result = parser.parse("index.html", "<div><span>Unclosed div");
-        assert!(result.is_ok());
-    }
-
-    // ── Integration tests ─────────────────────────────────────────────
-
-    #[test]
-    fn test_complex_raw_text_full_page() {
-        // Port of: 'should process a complex raw text'
-        let html = r#"<!DOCTYPE HTML><html dir="auto" lang="en"><head><meta charset="utf-8"><title>Test</title><style>html { margin: 0; }</style></head><body><app-shell></app-shell><script type="module" src="./index.js"></script></body></html>"#;
-        let (fragments, _) = parse_and_get_fragments(html);
-
-        // DOCTYPE + head + <body>, body_start, body content, body_end, </body></html>
-        assert!(fragments.len() >= 5);
-        assert!(
-            matches!(fragments[0].fragment.as_ref(), Some(Fragment::Raw(raw)) if
-                raw.value.contains("<!DOCTYPE HTML>") && raw.value.ends_with("<body>"))
-        );
-        assert!(
-            matches!(fragments[1].fragment.as_ref(), Some(Fragment::Signal(s)) if
-                s.value == "body_start" && s.raw)
-        );
-        assert!(
-            matches!(fragments[2].fragment.as_ref(), Some(Fragment::Raw(raw)) if
-                raw.value.contains("<app-shell>"))
-        );
-        assert!(
-            matches!(fragments[3].fragment.as_ref(), Some(Fragment::Signal(s)) if
-                s.value == "body_end" && s.raw)
-        );
-        assert!(
-            matches!(fragments[4].fragment.as_ref(), Some(Fragment::Raw(raw)) if
-                raw.value.contains("</body>") && raw.value.contains("</html>"))
-        );
-    }
-
-    #[test]
-    fn test_css_strategy_external_emits_link_tag() {
-        let mut parser = HtmlParser::new();
-        parser
-            .component_registry_mut()
-            .register_component("my-card", "<p><slot></slot></p>", Some("p { color: red; }"))
-            .ok();
-        parser.parse("index.html", "<my-card>Hello</my-card>").ok();
-        let records = parser.into_fragment_records();
-        let my_card = &records["my-card"].fragments;
-        let raw_text: String = my_card
-            .iter()
-            .filter_map(|f| match &f.fragment {
-                Some(Fragment::Raw(r)) => Some(r.value.as_str()),
-                _ => None,
-            })
-            .collect();
-        assert!(
-            raw_text.contains(r#"<link rel="stylesheet" href="./my-card.css">"#),
-            "Expected external <link> tag in: {}",
-            raw_text
-        );
-    }
-
-    #[test]
-    fn test_css_strategy_inline_emits_style_tag() {
-        let mut parser = HtmlParser::new();
-        parser.set_css_strategy(CssStrategy::Inline);
-        parser
-            .component_registry_mut()
-            .register_component("my-card", "<p><slot></slot></p>", Some("p { color: red; }"))
-            .ok();
-        parser.parse("index.html", "<my-card>Hello</my-card>").ok();
-        let records = parser.into_fragment_records();
-        let my_card = &records["my-card"].fragments;
-        let raw_text: String = my_card
-            .iter()
-            .filter_map(|f| match &f.fragment {
-                Some(Fragment::Raw(r)) => Some(r.value.as_str()),
-                _ => None,
-            })
-            .collect();
-        assert!(
-            raw_text.contains("<style>p { color: red; }</style>"),
-            "Expected inline <style> tag in: {}",
-            raw_text
-        );
-        assert!(
-            !raw_text.contains("<link"),
-            "Should not have <link> tag in inline mode: {}",
-            raw_text
-        );
-    }
-}
+mod tests;
