+++ crates/webui-parser/src/tests/mod.rs
@@ -1,0 +1,5 @@
+mod parse_attributes;
+mod parse_directives;
+mod parse_elements;
+mod parse_handlebars;
+mod parse_integration;
